DiseaseStatus, HealthStatus Controller 변경
DiseaseStatus, HealthStatus Controller에서 매핑주소 변경

+++ petpi/src/main/java/com/soon/petpi/service/PetService.java
@@ -1,11 +1,13 @@
 package com.soon.petpi.service;
 
+import com.soon.petpi.exception.type.NoPetError;
 import com.soon.petpi.model.dto.pet.PetCalenderResponse;
 import com.soon.petpi.model.dto.pet.PetRequest;
 import com.soon.petpi.model.dto.pet.PetResponse;
 import com.soon.petpi.model.entity.Pet;
 import com.soon.petpi.model.entity.User;
 import com.soon.petpi.repository.PetRepository;
+import com.soon.petpi.repository.UserRepository;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
@@ -24,64 +26,61 @@ public class PetService {
 
     private final PetRepository petRepository;
     private final FileStoreService fileStoreService;
+    private final UserRepository userRepository;
 
-    public Pet save(User user, PetRequest petRequest) throws IOException {
+    public Pet save(Long userIdx, PetRequest petRequest) throws IOException {
 
-        if (user == null) {
-            return null;
-        }
+        User petOwner = userRepository.findById(userIdx).orElse(null);
 
         Pet pet = petRequestToPet(petRequest);
-        pet.setUser(user);
+        pet.setUser(petOwner);
 
         petRepository.save(pet);
 
         return pet;
     }
 
-    public List<PetResponse> findAll(User user) {
+    public List<PetResponse> findAll(Long userIdx) {
 
-        if (user == null) {
-            return null;
-        }
-
-        Optional<List<Pet>> petsOptional = petRepository.findByUser(user);
+        Optional<List<Pet>> petsOptional = petRepository.findByUserIdx(userIdx);
         return petsOptional.map(pets ->
                 pets.stream().map(this::petToPetResponse)
                         .collect(Collectors.toList()))
                 .orElse(Collections.emptyList());
     }
 
-    public Pet findOne(Long petIdx) {
-        return petRepository.findById(petIdx).orElse(null);
+    public Pet findOne(Long petIdx, Long userIdx) {
+        Pet pet = petRepository.findByIdAndUserIdx(petIdx, userIdx).orElse(null);
+
+        if (pet == null) {
+            throw new NoPetError();
+        }
+        return pet;
     }
 
-    public Pet update(Long petIdx, PetRequest petRequest) throws IOException {
+    public Pet update(Long petIdx, Long userIdx, PetRequest petRequest) throws IOException {
 
-        Pet savedPet = findOne(petIdx);
+        Pet savedPet = findOne(petIdx, userIdx);
 
-        if (savedPet == null) {
-            return null;
-        }
+        log.info(savedPet.getPetIdx().toString());
 
-        if (savedPet.getPetImage() != null) {
+        if (savedPet.getPetImage() !=null && petRequest.getPetImage().equals(savedPet.getPetImage())) {
             fileStoreService.delete(savedPet.getPetImage());
         }
 
-        Pet updatePet = petRequestToPet(petRequest);
-        updatePet.setPetIdx(savedPet.getPetIdx());
-        updatePet.setUser(savedPet.getUser());
+        // Pet 객체 업데이트
+        savedPet.setPetName(petRequest.getPetName());
+        savedPet.setPetImage(fileStoreService.uploadFile(petRequest.getPetImage()).getStoreName());
+        savedPet.setPetGender(petRequest.getPetGender());
+        savedPet.setPetSpecies(petRequest.getPetSpecies());
+        savedPet.setPetBirthdate(petRequest.getPetBirthdate());
 
-        return petRepository.save(updatePet);
+        return petRepository.save(savedPet);
     }
 
-    public Boolean delete(Long petIdx) {
+    public Boolean delete(Long petIdx, Long userIdx) {
 
-        Pet deletePet = findOne(petIdx);
-
-        if(deletePet==null) {
-            return false;
-        }
+        Pet deletePet = findOne(petIdx, userIdx);
 
         if (deletePet.getPetImage() != null) {
             fileStoreService.delete(deletePet.getPetImage());
@@ -92,11 +91,11 @@ public class PetService {
         return true;
     }
 
-    public PetCalenderResponse readCalender(Long petIdx) {
-        Pet savedPet = petRepository.findByIdCalender(petIdx).orElse(null);
+    public PetCalenderResponse readCalender(Long petIdx, Long userIdx) {
+        Pet savedPet = petRepository.findByIdCalenderAndUserIdx(petIdx, userIdx).orElse(null);
 
         if (savedPet == null) {
-            return null;
+            throw new NoPetError();
         }
 
         return petToPetCalenderResponse(savedPet);
